Should fix the StaffChat.
REMEMBER: private Manager manager; not Manager manager;

If it doesn't work send me a DM on Discord and I will fix it as fast as I can. Just don't go messing with the code as it has a lot of dependencies. Also make sure the test server has either GroupManager or PermissionsEx, preferablly PermissionsEx.

<3 Hippo